+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## v2.0.2
+
+### Added
+
+- Add new String helpers:
+    + `replaceStrTo` Replace substr by start and finish indents
+
 ## v2.0.1
 
 ### Added